test(NODE-3688): reorganise tests
